Added roles. Admins, users, and managers see different stuff.

+++ app/views/department/department.blade.php
@@ -3,42 +3,51 @@
 @section('content')
 	<h2>Departments</h2>
 
-	<form id="newDepartment" action="index" method="post" class="pure-form pure-form-aligned">
-		<fieldset>
-			<legend>New Department</legend>
+	@if(Auth::user()->access_level === 'admin')
+		<form id="newDepartment" action="index" method="post" class="pure-form pure-form-aligned">
+			<fieldset>
+				<legend>New Department</legend>
 
-			<div class="pure-control-group">
-				<label for="addName">Name</label>
-				<input id="name" type="text" name="name">
-			</div>
+				<div class="pure-control-group">
+					<label for="addName">Name</label>
+					<input id="name" type="text" name="name">
+				</div>
 
-			<div class="pure-control-group">
-				<label for="addParentDept">Parent Department</label>
-				<select id="parentDept" name="parent_department" style="width: 200px;">
-					<option value=""></option>
-					@foreach($departments as $dept)
-						<option value="{{$dept->id}}">{{$dept->name}}</option>
-					@endforeach
-				</select>
-			</div>
+				<div class="pure-control-group">
+					<label for="addParentDept">Parent Department</label>
+					<select id="parentDept" name="parent_department" style="width: 200px;">
+						<option value=""></option>
+						@foreach($departments as $dept)
+							<option value="{{$dept->id}}">{{$dept->name}}</option>
+						@endforeach
+					</select>
+				</div>
 
-			<input type="hidden" name="is_active" value="1">
-			<button type="submit" class="pure-button pure-button-primary">Save</button>
-			<button type="reset" class="pure-button">Reset</button>
-		</fieldset>
-	</form>
+				<input type="hidden" name="is_active" value="1">
+				<button type="submit" class="pure-button pure-button-primary">Save</button>
+				<button type="reset" class="pure-button">Reset</button>
+			</fieldset>
+		</form>
+	@endif
 
 
 	<table class="pure-table pure-table-horizontal">
 		<thead>
-			<th>Name</th><th>Parent</th><th>Edit</th><th>Delete</th>
+			<th>Name</th>
+			<th>Parent</th>
+			@if(Auth::user()->access_level === 'admin')
+				<th>Edit</th>
+				<th>Delete</th>
+			@endif
 		</thead>
 		@foreach($departments as $dept)
 			<tr id="departmentRow{{$dept->id}}">
 				<td>{{$dept->name}}</td>
 				<td>{{ $dept->getParentDepartment() == null ? "" : $dept->getParentDepartment()->name }}</td>
-				<td><button onclick="editDept('{{$dept->id}}', '{{$dept->name}}', '{{$dept->parent_department}}')" class="pure-button">Edit</button></td>
-				<td><button onclick="deleteDept('{{$dept->id}}')" class="pure-button">Delete</button></td>
+				@if(Auth::user()->access_level === 'admin')
+					<td><button onclick="editDept('{{$dept->id}}', '{{$dept->name}}', '{{$dept->parent_department}}')" class="pure-button">Edit</button></td>
+					<td><button onclick="deleteDept('{{$dept->id}}')" class="pure-button">Delete</button></td>
+				@endif
 			</tr>
 		@endforeach
 	</table>
